# nodejs-express-skeleton/swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample document for backend template."
  version: "1.0.0"
  title: "Backend Template"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "petstore.swagger.io"
basePath: "/v2"
tags:
- name: "user"
  description: "Operations about user"
  externalDocs:
    description: "Find out more about our store"
    url: "http://swagger.io"
schemes:
- "https"
- "http"
paths:
  /search-users:
    post:
      tags:
      - "user"
      summary: "Search user"
      description: "For searching users in smart connect or DB"
      operationId: "searchUser"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: body
          name: body
          schema:
            type: object
            required:
              - keyword
              - page
            properties:
              keyword:
                type: string
              page:
                type: integer
      responses:
        '200':
          description: "Search result" 
        '400':
          description: "Not found" 
  /add-user:
    post:
      tags:
      - "user"
      summary: "Add user"
      description: "For add users to DB"
      operationId: "addUser"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: body
          name: body
          schema:
            $ref: "#/definitions/User"
      responses:
        '200':
          description: "Search result" 
        '400':
          description: "Not found" 
  /modify-user/{employee_id}:
    post:
      tags:
      - "user"
      summary: "Modify user"
      description: "For modifying users to DB"
      operationId: "modifyUser"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: body
          name: body
          schema:
            type: "object"
            properties:
              department:
                type: string
              employee_name:
                type: string
              employee_email:
                type: string
        - in: path
          name: employee_id
          required: true
          type: string
      responses:
        '200':
          description: "Successfully modified" 
        '400':
          description: "Not found" 
  /delete-user/{employee_id}:
    post:
      tags:
      - "user"
      summary: "Delete user"
      description: "For delete users to DB"
      operationId: "deleteUser"
      produces:
        - "application/json"
      parameters:
        - in: path
          name: employee_id
          required: true
          type: "string"
      responses:
        '200':
          description: "Successfully deleted" 
        '400':
          description: "Not found" 

definitions:
  User:
    type: "object"
    properties:
      department:
        type: string
      employee_id:
        type: string
      employee_name:
        type: string
      employee_email:
        type: string
    xml:
      name: "User"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"